Fixing bug in specs with logout properties
In current specs, "logout" property is placed as a child of "login". But this is wrong. Logout is a sibling of "login" and should be on the same level. This issue in the specs made a lot of problems and bugs in our project, as UAA was not able to read logout.redirect properties.