UAA writes cert bundle to a new location
Prior to this commit, the pre-start script used the
update-ca-certificates command to generate a cert bundle in a well known
location (/etc/ssl/certs/ca-certificates.crt).  Other jobs may also be
trying to call update-ca-certificates to update the same cert bundle.
Unfortunately, there's a period of time during the execution of
update-ca-certificates in which the certificate bundle is deleted from
the file system. UAA is not expecting the cert bundle to be missing and
errors out in unexpected ways.

Now, when UAA generates this bundle, it ensures the bundle is written to
a specially named file which cannot be removed by other jobs. This is
done via an undocumented flag. It's probably fine.

[#165159048]

Signed-off-by: Jeremy Morony <jmorony@pivotal.io>
Co-authored-by: Jeremy Morony <jmorony@pivotal.io>