Truststore creation is only done in the pre-start
Prior to this commit, the trust store was built in the pre-start script,
and then again in the UAA start script. The certificate bundle is
generated from a set of certificates provided via os-conf as well as
some certificates provided via `uaa.ca_certs`. Because the contents of
the certificate bundle are volatile, UAA attempted to read this file
more than once to ensure it had picked up all the neccessary certs.

EXCEPT. Turns out UAA doesn't need the certs provided by os-conf so this
behavior can be simplified.

[#165160019]

Signed-off-by: Andrew Wittrock <awittrock@pivotal.io>
Co-authored-by: Andrew Wittrock <awittrock@pivotal.io>