UAA process reads truststore from /var/vcap/data/*
Previously the pre-start script was writing to the jobs directory, when
in fact bosh jobs are supposed to store ephemeral data in the
/var/vcap/data directory: https://bosh.io/docs/vm-config/#storage

The previous logic was very complicated. The truststore was generated to
/var/vcap/data/uaa/cert-cache/cacerts and then symlinked into
/var/vcap/jobs/uaa/config/cacerts. As far as I can tell, this was
totally unnecessary. Instead of this whole symlink business we now just
put the truststore in /var/vcap/data/uaa/cert-cache/cacerts.

Previously we tested that the certs were correctly generated into the
truststore file, but we never tested whether that file actually was
being passed to the tomcat process. This commit also backfills a test
that verifies that the tomcat process actually gets passed the correct
truststore file.

Finally, we no longer mount /var/vcap/data/uaa/cert-cache as writable
into the bpm container, as the `uaa` script no longer writes to the cert
cache in any way.

[#165160019]