Add ability to set portHeader in tomcat's server.xml
If UAA receives a request from a proxy/load balancer where a port mapping is happening,
any redirect from UAA will pick the default port number (443 if https), which is the wrong
port where that proxy/load balancer is listening to. This commit updates the tomcat server
config file `server.xml` to get the port from `X-Forwarded-Port` if it is specified. This header
should be set by the proxy/load balancer.